requirements.txt
streamlit
wordcloud
pandas
pillow
matplotlib
openpyxl